+++ src/lib/env.ts
@@ -1,16 +1,50 @@
+import { AppEnvironmentLabel } from '@/types/enum'
 import type { AppEnvironment } from '@/types/app'
 
-export const getAppEnvironment = (): AppEnvironment => {
+export const getAppEnvironment = (path: string): AppEnvironment => {
   /*
-   * Load build information
+   * Gather build information
    *
-   * - mode: https://vitejs.dev/guide/env-and-mode#modes
-   * - commit: the Git commit available within CI/CD
-   * - time: the build time set within CI/CD as a frozen value
+   * Taken from environment variables and the current URL path to determine the context the app is operating within.
+   *
+   * Consists of:
+   * - label: local development, production, etc.
+   * - commit: the Git commit, set within CI/CD
+   * - time: the build time, set within CI/CD as a frozen value
    */
+  let label = AppEnvironmentLabel.LocalDevelopment
+
+  if (path.includes('/review')) {
+    label = AppEnvironmentLabel.ReviewApp
+  } else if (path.includes('/integration/')) {
+    label = AppEnvironmentLabel.Integration
+  } else if (path.includes('/prod/')) {
+    label = AppEnvironmentLabel.Production
+  }
+
   return {
-    mode: import.meta.env.MODE,
+    label: label,
     commit: import.meta.env.VITE_BUILD_HASH || null,
     time: import.meta.env.VITE_BUILD_TIME || null,
   }
+}
+
+export const getAppBorderClasses = (env: AppEnvironment): string[] => {
+  if (env.label === AppEnvironmentLabel.LocalDevelopment) {
+    return ['border-black', 'dark:border-white']
+  }
+
+  if (env.label === AppEnvironmentLabel.ReviewApp) {
+    return ['border-indigo-500']
+  }
+
+  if (env.label === AppEnvironmentLabel.Integration) {
+    return ['border-orange-500']
+  }
+
+  if (env.label === AppEnvironmentLabel.Production) {
+    return ['border-white', 'dark:border-black']
+  }
+
+  return []
 }